requirements.txt
Flask
openai
pdfplumber
python-dotenv
werkzeug
